+++ src/main/java/com/m1s/m1sserver/Repository/MemberCounselResultRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.MemberCounselResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberCounselResultRepository extends JpaRepository<MemberCounselResult, Integer> {
+public interface MemberCounselResultRepository extends JpaRepository<MemberCounselResult, Long> {
 }
