+++ src/main/java/com/m1s/m1sserver/Repository/MemberInterestRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.MemberInterest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberInterestRepository extends JpaRepository<MemberInterest, Integer> {
+public interface MemberInterestRepository extends JpaRepository<MemberInterest, Long> {
 }
